feat(forms): Override default configs.
